module products
CRUD

+++ src/api/categories/deleteCategories.js
@@ -1,16 +1,17 @@
 async function handleDeleteCate(dataId) {
-  try {
-    const deleteResp = await fetch(`http://localhost:3000/categories/${dataId}`, {
-      method: "DELETE",
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire("Deleted!", "Your file has been deleted.", "success");
+            fetch(`http://localhost:3000/categories/${dataId}`, {
+                method: "DELETE",
+            });
+        }
     });
-    if (deleteResp.ok) {
-      console.log("The categories has been successfully deleted");
-    } else {
-      console.error("Error during categories deletion.");
-    }
-  } catch (error) {
-    console.log("Error fetch categories data:", error);
-  }
 }
 
 export default handleDeleteCate;
